Replace : by . to do not have problems when convert output of make to table
Every time when I add a new item at Makefile, I get the output of command make
to put at vscode, replace the ": " by "|", put a | in begin and end of
all rows to be possible visualize the help at readme file.

When the comment of entry at Makefile have two :, I got problems makign
the replaces.

Would be good to make this automatic

Signed-off-by: Vitor Mattos <vitor@php.rio>